CDAP-14844 remove plugin endpoint usage
Removed usage of plugin endpoints since they are replaced by the
pipeline system service. Instead, moving related logic to the
configurePipeline() method and setting the output schema directly
there. Made the schema property nullable so that the schema is
fetched from underlying storage if it is not provided.
The 'Get Schema' button is configured to not pass schema to the
plugin.